refactor: extract the two grabbers (and lots of adjacent code) outside the main module
