+++ grabit.py
@@ -6,7 +6,17 @@
 # ]
 # ///
 import click
-from grabit_lib import VERSION, BaseGrabber, OutputFlags, OutputFormat, OutputFormatList, RenderFlags, grabbers, output
+from grabit_lib import (
+    VERSION,
+    BaseGrabber,
+    GrabitError,
+    OutputFlags,
+    OutputFormat,
+    OutputFormatList,
+    RenderFlags,
+    grabbers,
+    output,
+)
 
 
 @click.command()
@@ -93,24 +103,27 @@ def save(
     Download an URL, convert it to Markdown with specified options, and save it to a file.
     """
 
-    grabber = next((g for g in grabbers if g.can_handle(url)), BaseGrabber())
-    output_format_enums: OutputFormatList = OutputFormatList(output_formats)
+    try:
+        grabber = next((g for g in grabbers if g.can_handle(url)), BaseGrabber())
+        output_format_enums: OutputFormatList = OutputFormatList(output_formats)
 
-    render_flags = RenderFlags(
-        include_source=include_source,
-        include_title=include_title,
-        yaml_frontmatter=yaml_frontmatter,
-    )
-    output_flags = OutputFlags(
-        output_formats=output_format_enums,
-        create_domain_subdir=create_domain_subdir,
-        overwrite=overwrite,
-    )
+        render_flags = RenderFlags(
+            include_source=include_source,
+            include_title=include_title,
+            yaml_frontmatter=yaml_frontmatter,
+        )
+        output_flags = OutputFlags(
+            output_formats=output_format_enums,
+            create_domain_subdir=create_domain_subdir,
+            overwrite=overwrite,
+        )
 
-    title, outputs = grabber.grab(
-        url, user_agent, use_readability_js, fallback_title, render_flags, output_format_enums
-    )
-    output(title, outputs, url, output_flags)
+        title, outputs = grabber.grab(
+            url, user_agent, use_readability_js, fallback_title, render_flags, output_format_enums
+        )
+        output(title, outputs, url, output_flags)
+    except GrabitError as e:
+        raise click.ClickException(str(e))
 
 
 if __name__ == "__main__":
