+++ src/grabit_lib/core/__init__.py
@@ -1,3 +1,5 @@
+from grabit_lib.core.dtos import OutputFlags, RenderFlags
+from grabit_lib.core.misc import GrabitError
 from grabit_lib.core.output_format import OutputFormat, OutputFormatList
 
-__all__ = ["OutputFormat", "OutputFormatList"]
+__all__ = ["OutputFormat", "OutputFormatList", "RenderFlags", "OutputFlags", "GrabitError"]
