+++ src/grabit_lib/core/misc.py
@@ -1,0 +1,2 @@
+class GrabitError(Exception):
+    pass
