+++ src/grabit_lib/grabbers/__init__.py
@@ -1,0 +1,4 @@
+from grabit_lib.grabbers.base_grabber import BaseGrabber
+from grabit_lib.grabbers.reddit_grabber import RedditGrabber
+
+__all__ = ["BaseGrabber", "RedditGrabber"]
